# src/main/resources/static/openapi/YONDO_api.yaml
openapi: 3.0.3
info:
  title: KCC API
  description: |
    LoggingController에서 제공하는 사용자 관련 API 명세  
    (질문에서 주어진 7개 기능: 로그인, 비밀번호 초기화, 회원가입, 회원정보 수정, 토큰 재발급, 로그아웃 및 토큰 만료 시 기본 응답)
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: 로컬 서버 (개발용)
paths:
...
  /api/auth/sign-out:
    post:
      summary: "로그아웃"
      description: |
        * identifier(아이디)를 받아 로그아웃 처리
      operationId: signOut
      requestBody:
        required: true
        description: "로그아웃에 필요한 JSON 데이터"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignOutRequest"
            examples:
              normal:
                summary: "정상 로그아웃 예시"
                value:
                  identifier: "testUser"
      responses:
        "200":
          description: "로그아웃 성공"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "로그아웃 성공"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SignOutRequest:
      type: object
      required:
        - identifier
      properties:
        identifier:
          type: string
          example: "testUser"
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 401
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "에러 메시지 내용"
        code:
          type: string
          example: "T001"
      description: "에러 응답 공통 구조"
  responses:
    InternalServerError:
      description: "서버 내부 오류"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            internalError:
              summary: "서버 에러 예시"
              value:
                status: 500
                success: false
                message: "Internal server error"
                code: "E999"